requirements.txt
openai>=1.30.0
python-dotenv>=1.0.1
beautifulsoup4>=4.12.3
lxml>=5.2.2
requests>=2.32.3
PyYAML>=6.0.2
requests>=2.31.0
